Style fixes and behaviour fix
